feat: create RemoveProjectForm
